+++ src/commonMain/kotlin/kotools/types/number/NonZeroInt.kt
@@ -4,8 +4,7 @@ import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerializationException
 import kotools.types.Package
 import kotools.types.SinceKotoolsTypes
-import kotools.types.collection.NotEmptySet
-import kotools.types.collection.notEmptySetOf
+import kotools.types.range.NotEmptyRange
 import kotools.types.text.NotBlankString
 import kotools.types.text.toNotBlankString
 
@@ -16,28 +15,43 @@ public sealed interface NonZeroInt : AnyInt {
     /** Contains declarations for holding or building a [NonZeroInt]. */
     public companion object {
         /** The minimum value a [NonZeroInt] can have. */
+        @Deprecated(
+            "Use the negativeRange property instead.",
+            ReplaceWith("NonZeroInt.negativeRange.start.value")
+        )
         public val min: StrictlyNegativeInt by lazy(
-            StrictlyNegativeInt.Companion::min
+            StrictlyNegativeInt.range.start::value
         )
 
         /** The maximum value a [NonZeroInt] can have. */
+        @Deprecated(
+            "Use the positiveRange property instead.",
+            ReplaceWith("NonZeroInt.positiveRange.end.value")
+        )
         public val max: StrictlyPositiveInt by lazy(
-            StrictlyPositiveInt.Companion::max
+            StrictlyPositiveInt.range.end::value
+        )
+
+        /** The negative range of values a [NonZeroInt] can have. */
+        @SinceKotoolsTypes("4.2")
+        public val negativeRange: NotEmptyRange<StrictlyNegativeInt> by lazy(
+            StrictlyNegativeInt.Companion::range
+        )
+
+        /** The positive range of values a [NonZeroInt] can have. */
+        @SinceKotoolsTypes("4.2")
+        public val positiveRange: NotEmptyRange<StrictlyPositiveInt> by lazy(
+            StrictlyPositiveInt.Companion::range
         )
 
         /** Returns a random [NonZeroInt]. */
         @SinceKotoolsTypes("3.0")
-        public fun random(): NonZeroInt {
-            val ranges: NotEmptySet<IntRange> = notEmptySetOf(
-                min.toInt()..StrictlyNegativeInt.max.toInt(),
-                StrictlyPositiveInt.min.toInt()..max.toInt()
-            )
-            return ranges.toSet()
-                .random()
-                .random()
-                .toNonZeroInt()
-                .getOrThrow()
-        }
+        public fun random(): NonZeroInt = setOf(negativeRange, positiveRange)
+            .random()
+            .toIntRange()
+            .random()
+            .toNonZeroInt()
+            .getOrThrow()
     }
 
     @SinceKotoolsTypes("4.0")