+++ src/commonTest/kotlin/kotools/types/number/PositiveIntTest.kt
@@ -6,23 +6,42 @@ import kotlinx.serialization.SerializationException
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import kotools.types.Package
+import kotools.types.range.InclusiveBound
+import kotools.types.range.NotEmptyRange
 import kotools.types.shouldEqual
 import kotools.types.shouldHaveAMessage
 import kotools.types.shouldNotEqual
 import kotlin.test.Test
 import kotlin.test.assertFailsWith
+import kotlin.test.assertTrue
 
 class PositiveIntCompanionTest {
+    @Suppress("DEPRECATION")
     @Test
     fun min_should_equal_zero() {
         val result: ZeroInt = PositiveInt.min
         result shouldEqual ZeroInt
     }
 
+    @Suppress("DEPRECATION")
     @Test
     fun max_should_equal_the_maximum_value_of_Int() {
         val result: StrictlyPositiveInt = PositiveInt.max
         result.toInt() shouldEqual Int.MAX_VALUE
+    }
+
+    @Test
+    fun range_should_start_with_an_InclusiveBound_that_equals_zero() {
+        val range: NotEmptyRange<PositiveInt> = PositiveInt.range
+        assertTrue { range.start is InclusiveBound }
+        range.start.value shouldEqual ZeroInt
+    }
+
+    @Test
+    fun range_should_end_with_an_InclusiveBound_that_equals_the_maximum_value_of_Int() {
+        val range: NotEmptyRange<PositiveInt> = PositiveInt.range
+        assertTrue { range.end is InclusiveBound }
+        range.end.value.toInt() shouldEqual Int.MAX_VALUE
     }
 
     @Test