+++ src/commonTest/kotlin/kotools/types/number/StrictlyNegativeIntTest.kt
@@ -6,23 +6,44 @@ import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import kotools.types.Package
+import kotools.types.range.InclusiveBound
+import kotools.types.range.NotEmptyRange
 import kotools.types.shouldEqual
 import kotools.types.shouldHaveAMessage
 import kotools.types.shouldNotEqual
 import kotlin.test.Test
 import kotlin.test.assertFailsWith
+import kotlin.test.assertTrue
 
 class StrictlyNegativeIntCompanionTest {
+    @Suppress("DEPRECATION")
     @Test
     fun min_should_equal_the_minimum_value_of_Int() {
         val result: StrictlyNegativeInt = StrictlyNegativeInt.min
         result.toInt() shouldEqual Int.MIN_VALUE
     }
 
+    @Suppress("DEPRECATION")
     @Test
     fun max_should_equal_minus_one() {
         val result: StrictlyNegativeInt = StrictlyNegativeInt.max
         result.toInt() shouldEqual -1
+    }
+
+    @Test
+    fun range_should_start_with_an_InclusiveBound_that_equals_the_minimum_value_of_Int() {
+        val range: NotEmptyRange<StrictlyNegativeInt> =
+            StrictlyNegativeInt.range
+        assertTrue { range.start is InclusiveBound }
+        range.start.value.toInt() shouldEqual Int.MIN_VALUE
+    }
+
+    @Test
+    fun range_should_end_with_an_InclusiveBound_that_equals_minus_one() {
+        val range: NotEmptyRange<StrictlyNegativeInt> =
+            StrictlyNegativeInt.range
+        assertTrue { range.end is InclusiveBound }
+        range.end.value.toInt() shouldEqual -1
     }
 
     @Test