+++ changelog.md
@@ -22,8 +22,19 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-The `plus` operations for concatenating a `NotBlankString` with a `String` or a
-`Char` (issue [#53](https://github.com/kotools/types/issues/53)).
+- The `NotEmptyRange` and the `Bound` types representing a range of comparable
+  values that contain at least one value (issue
+  [#56](https://github.com/kotools/types/issues/56)).
+
+```kotlin
+val start: InclusiveBound<Int> = 1.toInclusiveBound()
+val end: ExclusiveBound<Int> = 42.toExclusiveBound()
+val range: NotEmptyRange<Int> = start..end // or end..start
+println(range) // [1;42[
+```
+
+- The `plus` operations for concatenating a `NotBlankString` with a `String` or
+  a `Char` (issue [#53](https://github.com/kotools/types/issues/53)).
 
 ```kotlin
 resultOf {
@@ -41,6 +52,17 @@ resultOf {
 }
 ```
 
+- The `NotEmptyRange` and the `Bound` types representing a range of comparable
+  values that contain at least one value (issue
+  [#56](https://github.com/kotools/types/issues/56)).
+
+```kotlin
+val start: InclusiveBound<Int> = 1.toInclusiveBound()
+val end: ExclusiveBound<Int> = 42.toExclusiveBound()
+val range: NotEmptyRange<Int> = start..end
+println(range) // [1;42[
+```
+
 ### Changed
 
 - Support for
@@ -56,6 +78,22 @@ resultOf {
 implementation("io.github.kotools:types:$version")
 // after
 implementation("org.kotools:types:$version")
+```
+
+### Deprecated
+
+The `min` and the `max` companion's properties of the `StrictlyPositiveInt`, the
+`StrictlyNegativeInt`, the `PositiveInt`, the `NegativeInt` and the `NonZeroInt`
+types (issue [#56](https://github.com/kotools/types/issues/56)).
+
+```kotlin
+var result: StrictlyPositiveInt
+// before
+result = StrictlyPositiveInt.min
+result = StrictlyPositiveInt.max
+// after
+result = StrictlyPositiveInt.range.start.value
+result = StrictlyPositiveInt.range.end.value
 ```
 
 ## 4.1.0